Split EditorVR into partial class files
